feat: reporting combined, preprocessing extracted
