+++ fairmlhealth/model_comparison.py
@@ -92,7 +92,7 @@ def compare_measures(test_data, target_data, protected_attr_data, models):
         pandas dataframe of fairness and performance measures for each model
     """
     warnings.warn(
-            "compare_measures will be deprecated in version 2." +
+            "compare_measures function will be deprecated in version 2.0" +
             " Use compare_models instead.", PendingDeprecationWarning
         )
     comp = FairCompare(test_data, target_data, protected_attr_data, models,
@@ -197,7 +197,7 @@ class FairCompare(ABC):
         if model_name not in self.preds.keys():
             msg += (" Name not found Available options include "
                    f"{list(self.preds.keys())}")
-            print(err)
+            print(msg)
             return pd.DataFrame()
         elif self.preds[model_name] is None:
             msg += (" No predictions present.")
@@ -209,11 +209,6 @@ class FairCompare(ABC):
                             self.probs[model_name], **kwargs)
         return res
 
-    def save_comparison(self, filepath):
-        dirpath = os.path.dirname(filepath)
-        if not os.path.exists(dirpath):
-            os.makedirs(dirpath)
-        dump(self, filepath)
 
     def __paused_validation(self):
         if not hasattr(self, "__pause_validation"):
